+++ backend/app/Http/Controllers/Admin/UserController.php
@@ -8,36 +8,57 @@ use App\Http\Requests\SaveUserRequest;
 use App\Models\User;
 use Exception;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Gate;
 
 
 class UserController extends Controller
 {
     public function index() : JsonResponse
     {
+        if(!Gate::any(['read-operators', 'read-admins'])) {
+            abort(403);
+        }
+
         $users = User::all();
         return response()->json(['users' => $users]);
     }
 
     public function store(SaveUserRequest $request) : JsonResponse
     {
+        if(!Gate::any(['create-operator', 'create-admin'])) {
+            abort(403);
+        }
+
         $user = (new CreateNewUser())->create($request->all());
         return response()->json(['user' => $user], 201);
     }
 
     public function show(int $id) : JsonResponse
     {
+        if(!Gate::any(['read-operator', 'read-admin'])) {
+            abort(403);
+        }
+
         $user = User::find($id);
         return response()->json(['user' => $user]);
     }
 
     public function update(SaveUserRequest $request, int $id) : JsonResponse
     {
+        if(!Gate::any(['update-operator', 'update-admin'])) {
+            abort(403);
+        }
+
         $user = User::findOrFail($id)->update($request->all());
         return response()->json(['updated' => $user]);
     }
 
     public function destroy(int $id) : JsonResponse
     {
+        if(!Gate::any(['delete-operator', 'delete-admin'])) {
+            abort(403);
+        }
+
         try {
             $deleted = User::findOrFail($id)->delete();
             $status = 200;
